inspector_model: give each module one concept
The files were named as the work arrived rather than by what they hold. `query.rs` held the `InspectedProgram` bundle after the queries it was named for were deleted; `snapshot.rs` held the wire types, using the word this codebase reserves for a single pane; `index.rs` and `links.rs` said neither which index nor whose links. Two modules also depended upward — `index.rs` called into `query.rs` for the shared traversal, and `snapshot.rs` reached into it for the payload builders.

Five modules, one concept each, no module depending on one below it: `walk` (the shared IR traversal), `span_index`, `name_binder`, `program` (`InspectedProgram` and its per-pane projections), and `wire` (the payload types, the schema version, the payload assembly, the pane-link projection). Tests move to the module that owns what they exercise. The public re-export set is unchanged, name for name.

A pure move: no behavior change, no wire change, no test dropped, 31 before and after.

With this the design doc's ratified list is empty, so it goes, and the doc now describes only what is in tree — the module layout among it, as fact rather than target. One citation went stale on the way: `provenance.md` named `stage::dense_edges`, a module path that had not existed for two commits, which `doc_refs` does not catch because a `::` path is not a source-path mention.
