+++ src/inspector_model/program.rs
@@ -1,29 +1,26 @@
-//! The [`InspectedProgram`] bundle — one compiled program's panes, indices and shared
-//! IR walk, assembled for the payload.
+//! The [`InspectedProgram`] bundle — one compiled program's panes and indices,
+//! assembled for the payload.
 //!
 //! [`InspectedProgram`] holds one [`PaneProjection`] per declared pane (that pane's IR
 //! tree, its `SourceProjection`, and the [`SpanIndex`] built over the pair), the
 //! pane-pair provenance maps, and the source-level [`NameBinderIndex`].
-//! [`build_payload`](InspectedProgram::build_payload) in `snapshot.rs` is its consumer.
+//! [`build_payload`](InspectedProgram::build_payload) in `wire.rs` is its consumer.
 //!
 //! There is no point-query layer: every static fact ships in the payload, and a
 //! positional question ("which node is at this position") is answered by the
 //! consumer over the shipped tables. See `src/inspector_model/design.md`, "The
 //! usage model".
 //!
-//! The shared IR walk lives here as well — [`predicate_children`],
-//! [`build_node_table`] and [`node_label`] — because the payload's per-pane
-//! node tables are built from it. No serde and no I/O: the wire
-//! types are `snapshot.rs`'s and the serialization is the `cambra-inspector`
-//! crate's.
+//! No serde and no I/O: the wire types and their builders are `wire.rs`'s, the
+//! shared IR walk is `walk.rs`'s, and the serialization is the
+//! `cambra-inspector` crate's.
 
 use crate::ccl::context::{CompiledProgram, Phase};
 use crate::ccl::panes::PANES;
 use crate::ccl::provenance::{NodeId, ProvenanceMap, SourceProjection};
-use crate::ccl::{Expr, FunKind, HistoryKind, Type, TypedBinding, TypedExprNode};
+use crate::ccl::{Expr, Type, TypedBinding};
 use crate::chl_parser::ast::{Module, Span};
 
-use super::snapshot::{IrChild, IrNode, RewriteInfo};
 use super::{Binding, NameBinderIndex, SpanIndex};
 
 /// The pane a binder's type is read from: the first fully-typed tree that is
@@ -344,253 +341,10 @@ impl<'a> InspectedProgram<'a> {
     }
 }
 
-/// The wire's discriminant for a type: which constructor it is, without its
-/// contents.
-///
-/// A rendered type is one string ([`crate::inspector_model`]'s design doc,
-/// "Types on the wire"), so a consumer that wants to branch — colour the
-/// function-typed nodes, filter to the refined ones — has nothing to branch on.
-/// This is that datum, and nothing more: it says which constructor sits at the
-/// top of the type, never what is inside it.
-///
-/// A function's kind is part of the discriminant, since the two render
-/// differently (`⇒` against `⤇`) and mean different things. An unpinned kind
-/// variable reads as a compute function, matching how `Display` renders it.
-pub(super) fn type_kind(ty: &Type) -> &'static str {
-    match ty {
-        Type::Base(_) => "base",
-        Type::UIntRange(_) => "range",
-        Type::Fun {
-            kind: FunKind::Data,
-            ..
-        } => "dataFun",
-        Type::Fun { .. } => "fun",
-        Type::Tuple(_) => "tuple",
-        Type::Record(_) => "record",
-        Type::Variant(..) => "variant",
-        Type::Refinement(..) => "refinement",
-        Type::Hole | Type::SharedHole(_) | Type::BoundedHole(_) => "hole",
-        Type::Infer(_) => "infer",
-        Type::DataSource(_) => "source",
-        Type::ChanDom(..) => "channel",
-        Type::Txn => "txn",
-        // A history is a mutable variable or a feed channel, and the two read
-        // differently enough that one discriminant would hide the distinction.
-        Type::History { kind, .. } => match kind {
-            HistoryKind::Overwrite => "mut",
-            HistoryKind::Append => "feed",
-        },
-    }
-}
-
-/// The predicates riding `ty` itself, as node ids, in the order
-/// [`predicate_children`] reaches them.
-///
-/// A node's `where.N` children cover every type slot it carries — its own type,
-/// an annotation, a `Cast` target, each binder's type — so a consumer holding
-/// them cannot tell which predicate refines *this node's* type. These are that
-/// subset, and because `predicate_children` walks the node's own type first,
-/// they are its leading children.
-pub(super) fn own_type_predicates(ty: &Type) -> Vec<u64> {
-    fn walk(t: &Type, out: &mut Vec<u64>) {
-        if let Type::Refinement(_, refinements) = t {
-            for r in refinements.iter() {
-                out.push(r.predicate.node_id().as_u64());
-            }
-        }
-        t.walk_children(|c| walk(c, out));
-    }
-    let mut out = Vec::new();
-    walk(ty, &mut out);
-    out
-}
-
-/// Every refinement predicate riding one of `expr`'s own type slots, paired with
-/// the wire label its child edge carries.
-///
-/// A predicate is a real expression tree with its own [`NodeId`]s, and the pane
-/// fold explains those ids: `collect_tree_ids`
-/// ([`crate::ccl::context`]) enumerates them, so they appear in every pane
-/// projection and as endpoints of every pane-pair map. A walk that stopped at
-/// `walk_children` would therefore ship links whose endpoints are absent from
-/// the pane they point into, which is what the wire validators call a dead
-/// endpoint. This is the descent that keeps the shipped node table and the
-/// shipped links over the same id domain.
-///
-/// The label is for display; the edge's `predicate` flag is what a consumer
-/// branches on to tell "this subtree lives inside a type" from "this subtree is
-/// an operand". Order is [`TypedExpr::walk_type_slots`] order, which is stable,
-/// so the labels are stable too, and a consumer can compare one node's predicate
-/// edges across panes.
-///
-/// Mirrors `collect_tree_ids`' type-slot descent, and must: a predicate that walk
-/// enumerates and this one does not is a node the fold explains and the table
-/// omits.
-pub(super) fn predicate_children(expr: &Expr) -> Vec<(String, &Expr)> {
-    fn from_ty<'t>(t: &'t Type, out: &mut Vec<&'t Expr>) {
-        if let Type::Refinement(_, refinements) = t {
-            // Every refinement's predicate rides the slot, so each is its own
-            // `where.N` child — the same per-member descent `collect_tree_ids`
-            // makes, which is what keeps the shipped node table and the shipped
-            // links over one id domain.
-            for r in refinements.iter() {
-                out.push(&r.predicate);
-            }
-        }
-        t.walk_children(|c| from_ty(c, out));
-    }
-
-    let mut roots = Vec::new();
-    expr.walk_type_slots(|t| from_ty(t, &mut roots));
-    roots
-        .into_iter()
-        .enumerate()
-        .map(|(i, p)| (format!("where.{i}"), p))
-        .collect()
-}
-
-/// Build one pane's node table against its `projection`, returning the root
-/// node's id and every node reachable from `expr` exactly once, in first-visit
-/// pre-order.
-///
-/// The single source-linking node builder: every pane's payload nodes go
-/// through this one shape, parameterized only by its `(Expr, SourceProjection)`
-/// pair.
-///
-/// A node reached from several places — a refinement predicate shared by
-/// several type slots — is emitted once and named by id from each place that
-/// reaches it, so nothing repeats and the walk terminates on a shared term. The
-/// pre-order is what makes the emitted array byte-reproducible.
-pub(super) fn build_node_table(expr: &Expr, projection: &SourceProjection) -> (u64, Vec<IrNode>) {
-    fn visit(
-        expr: &Expr,
-        projection: &SourceProjection,
-        visited: &mut std::collections::HashSet<NodeId>,
-        out: &mut Vec<IrNode>,
-    ) -> u64 {
-        let id = expr.node_id();
-        if !visited.insert(id) {
-            return id.as_u64();
-        }
-
-        let mut node = IrNode {
-            label: node_label(&expr.node),
-            node_id: id.as_u64(),
-            span: None,
-            rewritten: None,
-            ty: expr.ty.to_string(),
-            type_kind: type_kind(&expr.ty),
-            predicate_refs: own_type_predicates(&expr.ty),
-            children: Vec::new(),
-        };
-        if let Some(attr) = projection.get(&id) {
-            // The rewrite channel: a `Nature::Source` tag — the root of a
-            // lowered source expression — null-compresses and carries no wire
-            // tag; every other node carries `{via, nature, label}`. The
-            // validators guard that `"source"` never ships.
-            let tag = &attr.rewritten;
-            if !tag.nature.is_source() {
-                node.rewritten = Some(RewriteInfo {
-                    via: format!("{:?}", tag.via),
-                    nature: tag.nature.wire_str().to_string(),
-                    label: tag.label.to_string(),
-                });
-            }
-            // The spans channel: the node's primary (narrowest) source span, if
-            // any.
-            node.span = attr
-                .spans
-                .iter()
-                .min_by_key(|s| s.end.saturating_sub(s.start))
-                .copied();
-        }
-
-        // The entry claims its pre-order slot before its children are walked, so
-        // the array is ordered by first visit rather than by completion.
-        let slot = out.len();
-        out.push(node);
-
-        let mut children = Vec::new();
-        for (idx, child) in expr.child_exprs().into_iter().enumerate() {
-            children.push(IrChild {
-                edge: idx.to_string(),
-                id: visit(child, projection, visited, out),
-                predicate: false,
-            });
-        }
-        // Predicate subtrees come after the value children so a consumer that
-        // reads children positionally is unaffected; `predicate` is what marks
-        // them (see [`predicate_children`]).
-        for (edge, predicate) in predicate_children(expr) {
-            children.push(IrChild {
-                edge,
-                id: visit(predicate, projection, visited, out),
-                predicate: true,
-            });
-        }
-        out[slot].children = children;
-
-        id.as_u64()
-    }
-
-    let mut visited = std::collections::HashSet::new();
-    let mut nodes = Vec::new();
-    let root = visit(expr, projection, &mut visited, &mut nodes);
-    debug_assert_eq!(
-        nodes.len(),
-        visited.len(),
-        "the node table holds one entry per visited id"
-    );
-    (root, nodes)
-}
-
-/// A short kind label for a node, mirroring the symbolic vocabulary at a glance
-/// (`BinOp(+)`, `Lit(1)`, `Var(x)`, …) — a payload tree row's `label`.
-pub(super) fn node_label(node: &TypedExprNode) -> String {
-    use TypedExprNode::*;
-    match node {
-        Lit(l) => format!("Lit({l:?})"),
-        Var(n) => format!("Var({n})"),
-        Builtin(b) => format!("Builtin({b})"),
-        Apply { .. } => "Apply".to_string(),
-        Cast { .. } => "Cast".to_string(),
-        BinOp { op, .. } => format!("BinOp({op:?})"),
-        UnaryOp(op, _) => format!("UnaryOp({op:?})"),
-        Lambda { param, .. } => format!("Lambda({})", param.name),
-        Aggregate { kind, .. } => format!("Aggregate({kind:?})"),
-        Let { binding, .. } => format!("Let({})", binding.name),
-        List(_) => "List".to_string(),
-        Case { .. } => "Case".to_string(),
-        VariantCtor { tag, .. } => format!("VariantCtor(.{tag})"),
-        Transact { .. } => "Transact".to_string(),
-        LetRec { .. } => "LetRec".to_string(),
-        For { .. } => "For".to_string(),
-        MutWrite { name, .. } => format!("MutWrite({name})"),
-        // The declaring half of `:=`, named after its binder as `MutWrite` is
-        // after its target.
-        MutDecl { binding, .. } => format!("MutDecl({})", binding.name),
-        Tuple(_) => "Tuple".to_string(),
-        Proj(k) => format!("Proj({k:?})"),
-        Record(_) => "Record".to_string(),
-        Source(s) => format!("Source({s})"),
-        Compose(_) => "Compose".to_string(),
-        // Two operations, not one with a mode: a copairing lands on the
-        // operands' coproduct, a disjoint join on their shared domain. The
-        // inspector names them apart because they are apart.
-        Copair(_) => "Copair".to_string(),
-        DisjointJoin(_) => "DisjointJoin".to_string(),
-        ExprStmt { .. } => "ExprStmt".to_string(),
-        Feed { name, .. } => format!("Feed({name})"),
-        Define { name, .. } => format!("Define({name})"),
-        Begin { .. } => "Begin".to_string(),
-        Defer => "Defer".to_string(),
-        Error => "Error".to_string(),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::ccl::TypedExprNode;
     use crate::ccl::context::Phase;
     use crate::ccl::context::{GlobalContext, compile_program};
     use crate::ccl::provenance::Nature;
@@ -604,16 +358,6 @@ mod tests {
         let mut ctx = GlobalContext::default();
         let consumer: Box<dyn Consumer> = Box::new(|| {});
         compile_program(&mut ctx, code, consumer).expect("program compiles")
-    }
-
-    /// The span of the `n`-th (0-based) byte occurrence of `needle` in `code`.
-    fn nth_span(code: &str, needle: &str, n: usize) -> Span {
-        let start = code
-            .match_indices(needle)
-            .nth(n)
-            .unwrap_or_else(|| panic!("occurrence {n} of {needle:?} not found"))
-            .0;
-        Span::new(start, start + needle.len())
     }
 
     /// The specialization-wrapper `Let`s that `coalesce_generalized_let`
@@ -773,6 +517,9 @@ a
         let prog = compile(code);
         let snap = InspectedProgram::new(&prog);
 
+        // The def's binding site is its name's span, which is what the call's
+        // use pairs with — so take it from `definitions` rather than
+        // reconstructing it from byte offsets.
         let call = code.find("f(1, 2)").expect("call present");
         let call_span = Span::new(call, call + 1);
         let def_span = snap
@@ -814,64 +561,6 @@ a
         }
     }
 
-    // ------------------------------------------------------------------------
-    // Span↔CCL (source↔IR) mapping over the payload.
-    //
-    // These pin which source construct maps to which IR node, over two programs
-    // whose lowering produces synthetic wrapper chains: a `yield` generator and
-    // a `defer()`/`<<` feed pipeline. They mirror the manual web-validation
-    // examples (`cambra-inspector/examples/{generator_min,defer_min}.chl`) but
-    // inline the source so the flow is exercised without the front end.
-    //
-    // The question they ask is the consumer's — "what is at this span" — and it
-    // is asked the way a consumer asks it: over the shipped `(span, nodeId)`
-    // rows and the shipped tree.
-    // ------------------------------------------------------------------------
-
-    const GENERATOR_SRC: &str = "\
-def squared(xs):
-    for x in xs:
-        yield x * x
-
-max(squared([1, 2, 3, 4]))
-";
-
-    const DEFER_SRC: &str = "\
-readings = [1, 2, 3, 4]
-totals = defer()
-totals << sum(readings)
-for x in readings:
-    totals << x
-max(totals)
-";
-
-    /// The labels of the nodes `pane_id` indexes at `span`: every `(span,
-    /// nodeId)` row of that pane whose span covers the query, resolved to the
-    /// node carrying that id in that pane's node table. This is the consumer's
-    /// lookup, over the two shipped tables and nothing else.
-    ///
-    /// Panics if a row names a node the table does not hold — the invariant
-    /// `span_index_round_trips_with_projection` (`index.rs`) pins.
-    fn labels_at(payload: &InspectorPayload, pane_id: &str, span: Span) -> Vec<String> {
-        let pane = payload
-            .panes
-            .iter()
-            .find(|s| s.id == pane_id)
-            .unwrap_or_else(|| panic!("the payload ships a {pane_id} pane"));
-        pane.span_index
-            .iter()
-            .filter(|row| row.span.start <= span.start && span.end <= row.span.end)
-            .map(|row| {
-                pane.nodes
-                    .iter()
-                    .find(|n| n.node_id == row.node_id.as_u64())
-                    .unwrap_or_else(|| panic!("row node {:?} is in the table", row.node_id))
-                    .label
-                    .clone()
-            })
-            .collect()
-    }
-
     /// The distinct types the payload's scope rows join to the binder named
     /// `name`. A binder joins one type wherever it is visible, so a well-formed
     /// payload answers with a single entry.
@@ -887,171 +576,5 @@ max(totals)
             }
         }
         out
-    }
-
-    /// GENERATOR: the source constructs that map name the expected IR node.
-    /// `x * x` → the `Mul` BinOp; `max(...)` → `Aggregate(Max)`; the list
-    /// literals → their `Lit` nodes.
-    #[test]
-    fn generator_mapped_spans_resolve_to_expected_nodes() {
-        let prog = compile(GENERATOR_SRC);
-        let payload = InspectedProgram::new(&prog).build_payload("test");
-        let at = |span| labels_at(&payload, "post-inference", span);
-
-        // The `x * x` body → the arithmetic-mul BinOp (a mono clone of the
-        // generator body, which *does* carry the body span).
-        let mul = at(nth_span(GENERATOR_SRC, "x * x", 0));
-        assert!(
-            mul.iter().any(|l| l.contains("BinOp(Arithmetic(Mul))")),
-            "`x * x` → Mul BinOp; got {mul:?}"
-        );
-
-        // `max(squared(...))` → the Max aggregate.
-        let max = at(nth_span(GENERATOR_SRC, "max", 0));
-        assert!(
-            max.iter().any(|l| l.contains("Aggregate(Max)")),
-            "`max` → Max; got {max:?}"
-        );
-
-        // The list literals `1` and `2` map to their Lit nodes.
-        let lit1 = at(nth_span(GENERATOR_SRC, "1", 0));
-        assert!(
-            lit1.iter().any(|l| l.contains("Lit(Int(1))")),
-            "`1` → Lit; got {lit1:?}"
-        );
-        let lit2 = at(nth_span(GENERATOR_SRC, "2", 0));
-        assert!(
-            lit2.iter().any(|l| l.contains("Lit(Int(2))")),
-            "`2` → Lit; got {lit2:?}"
-        );
-
-        // The whole `[1, 2, 3, 4]` list literal (the argument of the
-        // monomorphized `squared(...)` call) maps to the `List` node. Span the
-        // elements, not the whole `[...]`: the `[` sits outside the lowered list
-        // span, so the elements' extent is what a row covers.
-        let list = at(nth_span(GENERATOR_SRC, "1, 2, 3, 4", 0));
-        assert!(
-            list.iter().any(|l| l.contains("List")),
-            "`[1, 2, 3, 4]` → List; got {list:?}"
-        );
-    }
-
-    /// DEFER: the source constructs that map name the expected IR node.
-    /// `sum(readings)` → `Aggregate(Sum)`; `max(totals)` → `Aggregate(Max)`; the
-    /// `totals` use in `max(totals)` → `Var(totals)`; the readings list literals
-    /// map.
-    #[test]
-    fn defer_mapped_spans_resolve_to_expected_nodes() {
-        let prog = compile(DEFER_SRC);
-        let payload = InspectedProgram::new(&prog).build_payload("test");
-        let at = |span| labels_at(&payload, "post-inference", span);
-
-        let sum = at(nth_span(DEFER_SRC, "sum", 0));
-        assert!(
-            sum.iter().any(|l| l.contains("Aggregate(Sum)")),
-            "`sum` → Sum; got {sum:?}"
-        );
-
-        let max = at(nth_span(DEFER_SRC, "max", 0));
-        assert!(
-            max.iter().any(|l| l.contains("Aggregate(Max)")),
-            "`max` → Max; got {max:?}"
-        );
-
-        // `totals` occurs 4×: the def (0), the two `<<` feeds (1, 2), and the
-        // `max(totals)` use (3). The last is the read whose span maps to Var.
-        let totals_use = at(nth_span(DEFER_SRC, "totals", 3));
-        assert!(
-            totals_use.iter().any(|l| l.contains("Var(totals)")),
-            "`totals` in `max(totals)` → Var(totals); got {totals_use:?}"
-        );
-
-        // The readings list literals map to Lit nodes.
-        let lit1 = at(nth_span(DEFER_SRC, "1", 0));
-        assert!(
-            lit1.iter().any(|l| l.contains("Lit(Int(1))")),
-            "`1` → Lit; got {lit1:?}"
-        );
-    }
-
-    /// DEFER: the **copaired fan-in** carries a source span.
-    ///
-    /// `Copair` is the node the `defer()`/`<<`/`for`-feed plumbing fans into,
-    /// tagged `via: Channelize, nature: Expansion`. It is indexed at the
-    /// `totals = defer()` statement — the declaration the feeds fan into — so a
-    /// consumer clicking the `defer()` site reaches it. Distinct from the feed
-    /// plumbing (`Lambda(__unused)`, the `Compose` over the feed body), tagged
-    /// `nature: Machinery`, which carries no span by design.
-    ///
-    /// The fan-in is a post-channelize artifact, absent from the post-inference
-    /// anchor, so the payload under test carries the post-channelize pane
-    /// alone.
-    #[test]
-    fn defer_coverage_maps_the_copaired_fan_in() {
-        let prog = compile(DEFER_SRC);
-        let panes = prog.materialize_panes();
-        let payload = InspectedProgram::from_parts(
-            "post-channelize",
-            &prog.source,
-            &prog.post_channelize_ir,
-            panes.projection("post-channelize").clone(),
-            &prog.source_ast,
-        )
-        .build_payload("test");
-        let at = |span| labels_at(&payload, "post-channelize", span);
-
-        // Copairing, not a disjoint join: the arms land on their coproduct, and
-        // nothing asserts the two feeds cover disjoint parts of one domain.
-        let defer_site = at(nth_span(DEFER_SRC, "defer()", 0));
-        assert!(
-            defer_site.iter().any(|l| l.contains("Copair")),
-            "the `defer()` declaration reaches the copaired fan-in; got {defer_site:?}"
-        );
-
-        // The fed value the user wrote still maps to its own aggregate, and the
-        // rest of the Part-A set maps at this pane too.
-        let feed = at(nth_span(DEFER_SRC, "sum", 0));
-        assert!(
-            feed.iter().any(|l| l.contains("Aggregate(Sum)")),
-            "`sum` → Sum; got {feed:?}"
-        );
-        let max = at(nth_span(DEFER_SRC, "max", 0));
-        assert!(
-            max.iter().any(|l| l.contains("Aggregate(Max)")),
-            "`max` → Max; got {max:?}"
-        );
-        let totals_use = at(nth_span(DEFER_SRC, "totals", 3));
-        assert!(
-            totals_use.iter().any(|l| l.contains("Var(totals)")),
-            "`totals` in `max(totals)` → Var(totals); got {totals_use:?}"
-        );
-        for digit in ["1", "2", "3", "4"] {
-            let lit = at(nth_span(DEFER_SRC, digit, 0));
-            assert!(
-                lit.iter()
-                    .any(|l| l.contains(&format!("Lit(Int({digit}))"))),
-                "the readings literal `{digit}` → Lit; got {lit:?}"
-            );
-        }
-
-        // Every node of the shipped table carries an attribution: the
-        // fully-folded rows leave no node absent from the projection, and a node
-        // the projection does not cover would ship neither a span nor a rewrite
-        // tag.
-        let pane = payload
-            .panes
-            .iter()
-            .find(|s| s.id == "post-channelize")
-            .expect("the payload ships the post-channelize pane");
-        let absent: Vec<&str> = pane
-            .nodes
-            .iter()
-            .filter(|n| n.span.is_none() && n.rewritten.is_none())
-            .map(|n| n.label.as_str())
-            .collect();
-        assert!(
-            absent.is_empty(),
-            "no defer node is left untagged (absent from projection); got {absent:?}"
-        );
     }
 }
